% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterClaraPAM}
\alias{clusterClaraPAM}
\title{Clara: Clustering algorithm for large datasets using PAM}
\usage{
clusterClaraPAM(sample)
}
\arguments{
\item{sample}{size <1000 (performance issues)}

\item{data}{dataframe}
}
\description{
Clara: Clustering algorithm for large datasets using PAM
}
\examples{
claraPAM(100)
}

