% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterKM}
\alias{clusterKM}
\title{KMean cluster over a dataset}
\usage{
clusterKM(sData, k, iter)
}
\arguments{
\item{sData}{subset of the data}

\item{k}{number of clusters}

\item{iter}{number of iterations}
}
\value{
kmean cluster
}
\description{
KMean cluster over a dataset
}
\examples{
cl<-clusterKM(sData,50,100)
gcl<-clusterKM(sData,20,100
}

