% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterValidation}
\alias{clusterValidation}
\title{Calculate cluster metrics
http://www.jstatsoft.org/v25/i04/paper}
\usage{
clusterValidation(data, cluster)
}
\arguments{
\item{data}{dataframe}

\item{cl}{kmean cluster}

\item{label}{of the cluster: E.g kmeans k: 50 between: 982683.05 withinss: 222461.94 totss: 1205145}
}
\value{
kmeans k: 50 between: 982683.05 withinss: 222461.94 totss: 1205145
}
\description{
Calculate cluster metrics
http://www.jstatsoft.org/v25/i04/paper
}
\examples{
labelKMean(cl)
}

