% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{getSubset}
\alias{getSubset}
\title{Fetch the data for a given image Descriptor (imgDescriptor) for a page range}
\usage{
getSubset(imgDescriptor, startPage, endPage)
}
\arguments{
\item{imgDescriptor}{in {SIFT, etc.}}

\item{startPage,}{start page (of the book) of the range (>=startPage)}

\item{endPage,}{end page (of the book) of the range (<endPage)}
}
\value{
matrix resulted from the split.
}
\description{
Fetch the data for a given image Descriptor (imgDescriptor) for a page range
}
\examples{
getSubset("SIFT")
getSubset(data[,2])
}

