% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{labelKMean}
\alias{labelKMean}
\title{Create a label of the kmean}
\usage{
labelKMean(cl)
}
\arguments{
\item{cl}{kmean cluster}
}
\value{
kmeans k: 50 between: 982683.05 withinss: 222461.94 totss: 1205145
}
\description{
Create a label of the kmean
}
\examples{
labelKMean(cl)
}

