% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{orchestrator}
\alias{orchestrator}
\title{Manage the complete process: get subset, cluster, validate, upload cluster to database}
\usage{
orchestrator()
}
\description{
Manage the complete process: get subset, cluster, validate, upload cluster to database
}
\examples{
orchestrator()
orchestrator()
}

