% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImageResize.R
\name{resizeImagesDB}
\alias{resizeImagesDB}
\title{Resize the patterns (images) of range of pages in a book}
\usage{
resizeImagesDB(startPage, endPage, x)
}
\arguments{
\item{startPage}{Number of the start page of the range in the book}

\item{endPage}{Number of the end page of the range in the book}

\item{x,}{height of the image}
}
\description{
Resize the patterns (images) of range of pages in a book
}
\examples{
resizeImagesPageDB(4,30)
}

