% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImageResize.R
\name{resizeImagesPageDB}
\alias{resizeImagesPageDB}
\title{Resize the patterns (images) of a page in a book}
\usage{
resizeImagesPageDB(page, x)
}
\arguments{
\item{page}{Number of the page in the book}

\item{x,}{height of the image}
}
\description{
Resize the patterns (images) of a page in a book
}
\examples{
resizeImagesPageDB(4,30)
}

