% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{splitCSVColumn}
\alias{splitCSVColumn}
\title{Split a comma separated column into a matrix.}
\usage{
splitCSVColumn(column)
}
\arguments{
\item{column}{or vector with the comma separated values.}
}
\value{
matrix resulted from the split.
}
\description{
Split a comma separated column into a matrix.
}
\examples{
splitCSVColumn("1,2,4,5")
splitCSVColumn(data[,2])
}

