% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{testCSVSQL}
\alias{testCSVSQL}
\title{Test Split a comma separated column into a matrix.}
\usage{
testCSVSQL(maxrows)
}
\arguments{
\item{maxrows}{Maximum numbers of rows to pull from mysql.}
}
\value{
matrix resulted from the split.
}
\description{
Test Split a comma separated column into a matrix.
}
\examples{
splitCSVColumn("1,2,4,5")
splitCSVColumn(data[,2])
}

