% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster.R
\name{uploadClusterDB}
\alias{uploadClusterDB}
\title{Upload a cluster into the database}
\usage{
uploadClusterDB(cluster, data, label, descriptor)
}
\arguments{
\item{cluster}{Vector of the cluster assigment for each point in the dataset}

\item{data}{dataframe}

\item{label}{of the cluster: E.g kmeans k: 50 between: 982683.05 withinss: 222461.94 totss: 1205145}
}
\description{
Upload a cluster into the database
}
\examples{
uploadClusterDB(l,data,"k mean k=10,descriptor)
uploadClusterDB(l,data,"k mean",descriptor)
}

